Move tail sampling to new package
