+++ modules/tail-sampling/src/main/scala/trace4cats/samplling/tail/RateTailSpanSampler.scala
@@ -1,11 +1,10 @@
-package io.janstenpickle.trace4cats.rate.sampling
+package trace4cats.samplling.tail
 
 import cats.effect.kernel.{Resource, Temporal}
 import cats.syntax.foldable._
 import cats.syntax.functor._
 import cats.{Applicative, Foldable, Monad, MonoidK}
 import trace4cats.model.{CompletedSpan, SampleDecision, TraceId}
-import io.janstenpickle.trace4cats.sampling.tail.{SampleDecisionStore, TailSpanSampler}
 import trace4cats.rate.TokenBucket
 
 import scala.concurrent.duration.FiniteDuration
