+++ modules/tail-sampling/src/main/scala/trace4cats/samplling/tail/SampleDecisionStore.scala
@@ -1,4 +1,4 @@
-package io.janstenpickle.trace4cats.sampling.tail
+package trace4cats.samplling.tail
 
 import trace4cats.model.{SampleDecision, TraceId}
